requirement.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.4.2
platformdirs==2.4.1
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.13.0
